+++ app/models/child.rb
@@ -251,7 +251,7 @@ class Child < UuidApplicationRecord
   def ne_hours(date, schedule_for_weekday)
     # TODO: this is super sloppy because this shouldn't be a service class
     # but we haven't refactored these to procedures yet
-    NebraskaHoursCalculator.new(
+    Nebraska::HoursCalculator.new(
       child: self,
       date: date,
       scope: :for_month
@@ -351,7 +351,7 @@ class Child < UuidApplicationRecord
   def nebraska_hours(date)
     # feature flag for using live algorithms rather than uploaded data
     if Rails.application.config.ff_ne_live_algorithms
-      NebraskaHoursCalculator.new(
+      Nebraska::HoursCalculator.new(
         child: self,
         date: date,
         scope: :for_month
@@ -367,7 +367,7 @@ class Child < UuidApplicationRecord
     if Rails.application.config.ff_ne_live_algorithms
       return 0 unless active_child_approval(date).hours
 
-      active_child_approval(date).hours - NebraskaHoursCalculator.new(
+      active_child_approval(date).hours - Nebraska::HoursCalculator.new(
         child: self,
         date: date,
         scope: nil
