+++ spec/models/child_spec.rb
@@ -178,13 +178,13 @@ RSpec.describe Child, type: :model do
     end
 
     describe '#nebraska_hours' do
-      let(:calculator_class) { class_double(NebraskaHoursCalculator, new: nil) }
-      let(:calculator_instance) { instance_double(NebraskaHoursCalculator, call: nil) }
+      let(:calculator_class) { class_double(Nebraska::HoursCalculator, new: nil) }
+      let(:calculator_instance) { instance_double(Nebraska::HoursCalculator, call: nil) }
 
       context 'when using live algorithms' do
-        it 'calls the NebraskaHoursCalculator service' do
+        it 'calls the Nebraska::HoursCalculator service' do
           allow(Rails.application.config).to receive(:ff_ne_live_algorithms).and_return(true)
-          allow(NebraskaHoursCalculator)
+          allow(Nebraska::HoursCalculator)
             .to receive(:new).with(child: child,
                                    date: Time.current.to_date,
                                    scope: :for_month).and_return(calculator_instance)
@@ -194,7 +194,7 @@ RSpec.describe Child, type: :model do
       end
 
       context 'when using temporary dashboard values' do
-        it 'does not call the NebraskaHoursCalculator service' do
+        it 'does not call the Nebraska::HoursCalculator service' do
           allow(Rails.application.config).to receive(:ff_ne_live_algorithms).and_return(false)
           child.nebraska_hours(Time.current.to_date)
           expect(calculator_class).not_to have_received(:new)
